refactor requests dir
