+++ api/app/Http/Requests/Language/LanguageStoreRequest.php
@@ -1,0 +1,30 @@
+<?php
+
+namespace App\Http\Requests\Language;
+
+use Illuminate\Foundation\Http\FormRequest;
+
+class LanguageStoreRequest extends FormRequest
+{
+    /**
+     * Determine if the user is authorized to make this request.
+     *
+     * @return bool
+     */
+    public function authorize()
+    {
+        return false;
+    }
+
+    /**
+     * Get the validation rules that apply to the request.
+     *
+     * @return array
+     */
+    public function rules()
+    {
+        return [
+            //
+        ];
+    }
+}
